requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
websockets>=11.0.0
aiohttp>=3.8.0
orjson>=3.9.0
networkx>=3.0
asyncio-mqtt>=0.16.0
python-telegram-bot>=20.0